@W-16361430 Bypass Kerberos auth if TLS auth is successful
TLS authentication and Kerberos authentication can both be configured and by default both particiate in connection negotiation. Introduce a new configuration parameter that allows TLS authentication to be sufficient for authentication if successful.

The client and wire protocol are not changed. The client may send a preamble asking for either SIMPLE or KERBEROS authentication, depending on how it is configured. It will also send the first SASL handshake message if KERBEROS. If configured to initiate TLS it will first complete the TLS handshake before sending connection preamble and initial SASL negotiation packet. On the server we process the TLS handshake and, if accepted, then either 1) accept the offered SIMPLE authentication or 2) terminate KERBEROS authentication and force the client to downgrade to SIMPLE, disregarding any kerberos credentials.